+++ server/test/integration/dbTest/getHistoricAlertsByAlertApiKeyTest.js
@@ -3,10 +3,9 @@ const { expect } = require('chai')
 const { afterEach, beforeEach, describe, it } = require('mocha')
 
 // In-house dependencies
-const { CHATBOT_STATE, ALERT_TYPE, factories, helpers } = require('brave-alert-lib')
+const { CHATBOT_STATE, factories, helpers } = require('brave-alert-lib')
 const db = require('../../../db/db')
-const SessionState = require('../../../SessionState')
-const { buttonDBFactory } = require('../../testingHelpers')
+const { buttonDBFactory, sessionDBFactory } = require('../../testingHelpers')
 
 describe('db.js integration tests: getHistoricAlertsByAlertApiKey', () => {
   describe('if there are no clients with the given Alert API Key', () => {
@@ -19,7 +18,10 @@ describe('db.js integration tests: getHistoricAlertsByAlertApiKey', () => {
         buttonId: '51b8be5678bf5ade9bf6a5958b2a4a45',
         clientId: client.id,
       })
-      await db.createSession(client.id, button.buttonId, 'unit', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
+      await sessionDBFactory(db, {
+        clientId: client.id,
+        buttonId: button.buttonId,
+      })
     })
 
     afterEach(async () => {
@@ -45,7 +47,10 @@ describe('db.js integration tests: getHistoricAlertsByAlertApiKey', () => {
         buttonId: '51b8be5678bf5ade9bf6a5958b2a4a45',
         clientId: client.id,
       })
-      await db.createSession(client.id, button.buttonId, 'unit', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
+      await sessionDBFactory(db, {
+        clientId: client.id,
+        buttonId: button.buttonId,
+      })
 
       // Insert a single client with no sessions that matches the Alert API Key that we ask for
       this.alertApiKey = 'alertApiKey'
@@ -82,34 +87,18 @@ describe('db.js integration tests: getHistoricAlertsByAlertApiKey', () => {
       })
 
       // Insert a single session for that API key
-      this.incidentType = ALERT_TYPE.BUTTONS_URGENT
+      this.incidentType = 'Cat 0'
       this.numPresses = 6
       this.respondedAt = new Date('2021-01-20T06:20:19.000Z')
-      this.session = await db.createSession(
-        client.id,
-        this.button.buttonId,
-        this.button.displayName,
-        'phoneNumber',
-        this.numPresses,
-        95,
-        this.respondedAt,
-      )
-      await db.saveSession(
-        new SessionState(
-          this.session.id,
-          this.session.clientId,
-          this.session.buttonId,
-          this.session.unit,
-          this.session.phoneNumber,
-          CHATBOT_STATE.WAITING_FOR_CATEGORY,
-          this.numPresses,
-          this.session.createdAt,
-          new Date(),
-          this.incidentType,
-          this.session.buttonBatteryLevel,
-          this.respondedAt,
-        ),
-      )
+      this.session = await sessionDBFactory(db, {
+        clientId: client.id,
+        buttonId: this.button.buttonId,
+        unit: this.button.displayName,
+        numPresses: this.numPresses,
+        respondedAt: this.respondedAt,
+        state: CHATBOT_STATE.WAITING_FOR_CATEGORY,
+        incidentType: this.incidentType,
+      })
     })
 
     afterEach(async () => {
@@ -153,11 +142,11 @@ describe('db.js integration tests: getHistoricAlertsByAlertApiKey', () => {
         displayName: 'unit2',
         buttonSerialNumber: 'button2',
       })
-      this.session1 = await db.createSession(client.id, button1.id, 'unit1', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
-      this.session2 = await db.createSession(client.id, button2.id, 'unit2', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
-      this.session3 = await db.createSession(client.id, button2.id, 'unit2', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
-      this.session4 = await db.createSession(client.id, button1.id, 'unit1', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
-      this.session5 = await db.createSession(client.id, button2.id, 'unit2', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
+      this.session1 = await sessionDBFactory(db, { clientId: client.id, buttonId: button1.id })
+      this.session2 = await sessionDBFactory(db, { clientId: client.id, buttonId: button2.id })
+      this.session3 = await sessionDBFactory(db, { clientId: client.id, buttonId: button2.id })
+      this.session4 = await sessionDBFactory(db, { clientId: client.id, buttonId: button1.id })
+      this.session5 = await sessionDBFactory(db, { clientId: client.id, buttonId: button2.id })
     })
 
     afterEach(async () => {
@@ -194,12 +183,12 @@ describe('db.js integration tests: getHistoricAlertsByAlertApiKey', () => {
         displayName: 'unit2',
         buttonSerialNumber: 'button2',
       })
-      this.session1 = await db.createSession(client.id, button1.id, 'unit1', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
-      this.session2 = await db.createSession(client.id, button2.id, 'unit2', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
+      this.session1 = await sessionDBFactory(db, { clientId: client.id, buttonId: button1.id })
+      this.session2 = await sessionDBFactory(db, { clientId: client.id, buttonId: button2.id })
       await helpers.sleep(2000) // Couldn't get around doing this because there is a trigger on updated_at to keep it correct
-      this.session3 = await db.createSession(client.id, button2.id, 'unit2', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
-      this.session4 = await db.createSession(client.id, button1.id, 'unit1', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
-      this.session5 = await db.createSession(client.id, button2.id, 'unit2', 'phoneNumber', 5, 95, new Date('2021-01-20T06:20:19.000Z'))
+      this.session3 = await sessionDBFactory(db, { clientId: client.id, buttonId: button2.id })
+      this.session4 = await sessionDBFactory(db, { clientId: client.id, buttonId: button1.id })
+      this.session5 = await sessionDBFactory(db, { clientId: client.id, buttonId: button2.id })
     })
 
     afterEach(async () => {
@@ -230,15 +219,11 @@ describe('db.js integration tests: getHistoricAlertsByAlertApiKey', () => {
         displayName: 'unit1',
         clientId: client.id,
       })
-      this.session = await db.createSession(
-        client.id,
-        button.buttonId,
-        button.displayName,
-        'phoneNumber',
-        '1',
-        95,
-        new Date('2021-01-20T06:20:19.000Z'),
-      )
+      this.session = await sessionDBFactory(db, {
+        clientId: client.id,
+        buttonId: button.buttonId,
+        unit: button.displayName,
+      })
     })
 
     afterEach(async () => {