+++ server/test/testingHelpers.js
@@ -1,23 +1,39 @@
 // In-house dependencies
-const { factories } = require('brave-alert-lib')
-const SessionState = require('../SessionState')
+const { factories, CHATBOT_STATE } = require('brave-alert-lib')
+const Session = require('../Session')
 const Hub = require('../Hub')
 
-function createTestSessionState() {
-  return new SessionState(
-    'ca6e85b1-0a8c-4e1a-8d1e-7a35f838d7bc',
-    'fakeClientId',
-    'fakeButtonId',
-    'fakeUnit',
-    'fakePhone',
-    'fakeState',
-    'fakeNumPresses',
-    'fakeCreatedAt',
-    'fakeUpdatedAt',
-    '1',
-    null,
-    new Date('2000-06-06T00:53:53.000Z'),
+function sessionFactory(overrides = {}) {
+  return new Session(
+    overrides.id !== undefined ? overrides.id : 'ca6e85b1-0a8c-4e1a-8d1e-7a35f838d7bc',
+    overrides.clientId !== undefined ? overrides.clientId : 'fakeClientId',
+    overrides.buttonId !== undefined ? overrides.buttonId : 'fakeButtonId',
+    overrides.unit !== undefined ? overrides.unit : 'fakeUnit',
+    overrides.phoneNumber !== undefined ? overrides.phoneNumber : '18885554444',
+    overrides.state !== undefined ? overrides.state : CHATBOT_STATE.COMPLETED,
+    overrides.numPresses !== undefined ? overrides.numPresses : 1,
+    overrides.incidentType !== undefined ? overrides.incidentType : '1',
+    overrides.buttonBatteryLevel !== undefined ? overrides.buttonBatteryLevel : 100,
+    overrides.respondedAt !== undefined ? overrides.respondedAt : new Date('2000-06-06T00:53:53.000Z'),
+    overrides.createdAt !== undefined ? overrides.createdAt : 'fakeCreatedAt',
+    overrides.updatedAt !== undefined ? overrides.updatedAt : 'fakeUpdatedAt',
   )
+}
+
+async function sessionDBFactory(db, overrides = {}) {
+  const session = await db.createSession(
+    overrides.clientId !== undefined ? overrides.clientId : 'fakeClientId',
+    overrides.buttonId !== undefined ? overrides.buttonId : 'fakeButtonId',
+    overrides.unit !== undefined ? overrides.unit : 'fakeUnit',
+    overrides.phoneNumber !== undefined ? overrides.phoneNumber : '+18885554444',
+    overrides.state !== undefined ? overrides.state : CHATBOT_STATE.STARTED,
+    overrides.numPresses !== undefined ? overrides.numPresses : 1,
+    overrides.incidentType !== undefined ? overrides.incidentType : null,
+    overrides.buttonBatteryLevel !== undefined ? overrides.buttonBatteryLevel : null,
+    overrides.respondedAt !== undefined ? overrides.respondedAt : new Date('2000-06-06T00:53:53.000Z'),
+  )
+
+  return session
 }
 
 async function buttonDBFactory(db, overrides = {}) {
@@ -50,7 +66,8 @@ function hubFactory(overrides = {}) {
 }
 
 module.exports = {
-  createTestSessionState,
   buttonDBFactory,
   hubFactory,
+  sessionDBFactory,
+  sessionFactory,
 }