+++ server/test/unit/SessionStateTest.js
@@ -4,23 +4,16 @@ const expect = chai.expect
 const beforeEach = require('mocha').beforeEach
 const describe = require('mocha').describe
 const it = require('mocha').it
-const CHATBOT_STATE = require('brave-alert-lib').CHATBOT_STATE
 
-const SessionState = require('../../SessionState.js')
+const { sessionFactory } = require('../testingHelpers')
 
-describe('SessionState class', () => {
-  const sessionId = '12345'
-  const clientId = '67890'
-  const buttonId = '12345'
-  const unit = '1'
-  const phoneNumber = '+14206666969'
-  const createdAt = Date()
-  const updatedAt = Date()
-
+describe('Session class', () => {
   let state
 
   beforeEach(() => {
-    state = new SessionState(sessionId, clientId, buttonId, unit, phoneNumber, CHATBOT_STATE.STARTED, 1, createdAt, updatedAt, null)
+    state = sessionFactory({
+      numPresses: 1,
+    })
   })
 
   it('should start off with 1 button press', () => {